check file handling
With large datasets the checking of the file (counting lines) can take a few
seconds. This creates a window where a user could enter a title and click add
dataset, which lands them on the datasets page with a dataset that appears to
have no ids in it.

To prevent this there are new actions (START_CHECK_FILE and END_CHECK_FILE)
which allow the AddDataset component to know when a file is being checked so
the Add Dataset and Select Tweet ID file buttons can be disabled.