Initial Changes
The scalable-hydration branch is a series of changes to address #15.

These initial changes center on gutting the hydration in the main process for
the moment, and also normalizing all the Electron IPC messages so that
they reuse the names from the Redux actions. Hopefully this will make it
easier to trace the code, and catch errors.