Autosave
This commit gets autosave working in the main process again. It also
removes the redux-localstorage which came from the boilerplate. I wanted
to have a bit more control over how the saved storage is passed off to
the store, for example turning off any active downloads, since they
would have stopped with the application.

This commit also removes some dead files that were not cleaned up during
the port to the new boilerplate.