Rate Limiting
When we receive a rate limit error look up the reset time at Twitter and
use it. Previously we were relying on the fact that we would be able to
stop before we were rate limited. But this is not always the case when
multiple hydration jobs are running, or when the app is started up while
it has been rate limited.